+++ bciers/apps/reporting/src/app/components/taskList/headerElementFactory.ts
@@ -1,11 +1,13 @@
 import { TaskListElement } from "@bciers/components/navigation/reportingTaskList/types";
-import { HeaderStep } from "./types";
+import { HeaderStep, TaskListPageFactoryContext } from "./types";
 
 /**
  * For each header step, the tasklist can supply a root element to match the designs
  */
 export const headerElementFactories: {
-  [Step in HeaderStep]?: (context: any) => TaskListElement;
+  [Step in HeaderStep]?: (
+    context: TaskListPageFactoryContext,
+  ) => TaskListElement;
 } = {
   [HeaderStep.OperationInformation]: () => ({
     type: "Section",
@@ -26,7 +28,7 @@ export const headerElementFactories: {
 
 export function headerElementFactory(
   step: HeaderStep,
-  context: any,
+  context: TaskListPageFactoryContext,
 ): TaskListElement | undefined {
   const factory = headerElementFactories[step];
   return factory && factory(context);
