+++ bciers/apps/reporting/src/app/components/taskList/navigationInformation.ts
@@ -72,8 +72,6 @@ export async function getNavigationInformation(
 
   const pages = flowData[step] as ReportingPage[];
 
-  console.log(pages);
-
   // build pages from the factories, skipping the ones
   const tasklistPages = (
     await Promise.all(
@@ -88,8 +86,6 @@ export async function getNavigationInformation(
       }),
     )
   ).filter((p) => !p.extraOptions?.skip);
-
-  console.log(tasklistPages);
 
   const [taskListHeaderPages, taskListNonHeaderPages] =
     splitHeaderElements(tasklistPages);
