+++ bciers/apps/reporting/src/app/components/taskList/taskListPages/3_additionalInformation.ts
@@ -1,4 +1,4 @@
-import { ReportingPage, TaskListPageFactory } from "./types";
+import { ReportingPage, TaskListPageFactory } from "../types";
 
 export const additionalInformationPageFactories: {
   [Page in ReportingPage]?: TaskListPageFactory;
