+++ bciers/apps/reporting/src/app/components/taskList/taskListPages/4_complianceSummary.ts
@@ -1,4 +1,4 @@
-import { ReportingPage, TaskListPageFactory } from "./types";
+import { ReportingPage, TaskListPageFactory } from "../types";
 
 export const complianceSummaryPageFactories: {
   [Page in ReportingPage]?: TaskListPageFactory;
